requirements.txt
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
factories==1.2.0
factory-boy==3.2.1
Faker==15.3.4
fasteners==0.14.1
future==0.18.2
gunicorn==21.2.0
httpie==3.2.2
httplib2==0.14.0



